+++ src/nmea/sentences/mod.rs
@@ -2,6 +2,8 @@
 
 #[cfg(feature = "apb")]
 mod apb;
+#[cfg(feature = "bod")]
+mod bod;
 #[cfg(feature = "bwc")]
 mod bwc;
 #[cfg(feature = "bwr")]
@@ -54,6 +56,8 @@ mod pskpdpt;
 mod rmb;
 #[cfg(feature = "rmc")]
 mod rmc;
+#[cfg(feature = "rsd")]
+mod rsd;
 #[cfg(feature = "rot")]
 mod rot;
 #[cfg(feature = "rpm")]
@@ -87,6 +91,8 @@ mod zda;
 
 #[cfg(feature = "apb")]
 pub use apb::*;
+#[cfg(feature = "bod")]
+pub use bod::*;
 #[cfg(feature = "bwc")]
 pub use bwc::*;
 #[cfg(feature = "bwr")]
@@ -145,6 +151,8 @@ pub use rot::*;
 pub use rpm::*;
 #[cfg(feature = "rsa")]
 pub use rsa::*;
+#[cfg(feature = "rsd")]
+pub use rsd::*;
 #[cfg(feature = "ths")]
 pub use ths::*;
 #[cfg(feature = "txt")]